Typos, Grammar, More Content, and Hair Pulling Out
Gotta figure out a way to get those blasted lines to indent for me. I tried hitting it with the kitchen sink, to no avail, so tomorrow... the entire house.

Also, I _TOTALLY_ misspoke and had **not** manage to fix the Section Collapse issue... Because it's related to having commented out Ads. So when I'm done with all this, that will resolve itself! 👍 😏 _(at least I correctly identified it as user-error!)_