Sorted out the indentation woes! 😄
Visually functional.... but whether it's "properly done" is another matter entirely! 

Recycled the "Quote" box to work as a sort of "information legend" to encapsulate that section and make it appear more standalone. _(Wasn't needed to accomplish what I wanted, so it can always be removed if deemed unsightly and/or confusing.)_
 
Note: I plan to utilize this kludge for the entire Tag list, for each entry in order to indent those as well.... Just as a warning! 😊